pyproject.toml
[tool.poetry]
name = "database-configurations"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "traceitdb"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
Faker = "^15.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
